this is really not "nisp", contains org-mode related stuff
Pushing to github now because stuff depends on this file. I think after
this you can clone this repo and use everything in it. You still cannot
easily use just one part of it yet. Most things still depend on things they should not depend on (generally things that are too monolithic). I'm working on that.