Add nisp-file-name-buffer-directory
This takes a filename and optionally an extension and buffer and
produces a string with the filename appended to the current directory of
the buffer. If you think this is confusing, so did I when I first saw it.

As I rewrite I'll probably rewrite this function again to express itself
a little more clearly.