Add nisp-slime-toggle-trace-fdefinition-no-query
Toggle tracing on the common lisp side.

This is a modified version of `slime-toggle-trace-fdefinition'
that does not prompt for which function to trace, it just does it.

No prompt means toggling trace on and off of a function using a
single key press is much smoother. If the wrong thing is toggled,
its easy enough to re-invoke the function.

Please note that the optional parameter using-context-p is set to
t with no option to override. If this is a problem, I will add it
back in but with it defaulting to t instead of nil.