new macro: save-point-and-goto
This saves the point passed in as an argument and evaluates BODY with
point at that location. If point as passed to the macro is nil point
stays where it is, but its location is still saved so the BODY can
modify where point is and have it restored on normal exit.